backend-flask/requirements.txt
python-dotenv
flask
flask-cors

#Honeycomb

opentelemetry-api 
opentelemetry-sdk 
opentelemetry-exporter-otlp-proto-http 
opentelemetry-instrumentation-flask 
opentelemetry-instrumentation-requests
opentelemetry.exporter.otlp.proto.grpc

#X-Ray
aws-xray-sdk